Redefinition of pdb_t
PDBs were formerly implemented as closed lists. Worse, the definition of
npancake_t contained information relevant for ranking abstract states.

This changeset underwent a major modification. PDBs are implemented now
as such, and they contain all the required information for ranking
permutations and storing them. The type constraint pdb_type has been
extended so that the underlying type has to provide the permutation to
rank (and this makes libpdb of use only in permutation state spaces,
which is okay for my purposes).

The resulting definition of the underlying type, e.g., npancake_t is now
much easier and lighter.

This changeset contains unit tests for testing PDBs