Heavy-cost abstract states in the N-Pancake
This changeset contains a unit test for verifying that the descendants
of abstract states in the heavy-cost variant of the N-Pancake are
correctly generated, resulting in other abstract states generated with a
cost which, when being known is correct and, otherwise, it equals the
default cost.

The definition of npancake_t has been extended to include the concept of
default_cost which must be given during the initialization. It is the
value of an operator when the first disc immediately below the spatula
has been abstracted away, so that it is unknown. In a real application,
this should be equal to the radius of the smaller disc abstracted away.