pdb_t is now defined over node_t<pdb_type T>
PDBs use items of type node_t<T> for any type T satisfying the pdb_type
constraint. The definition of the template has changed to enforce that
items are of this type. Consequently, some unit tests were modified.

The result is now much more clear and easier to understand and follow!