Consistent use of pdbval_t
In general, the PDBs built here should allow the specification of any
type. Right now, only uint8_t is acknowledged and it is given indeed as
a type defintion named pdbval_t

To better prepare for modifications in the future, if any, uint8_t is
never used across the library (or tests, or domain specs). Instead
pdbval_t is consistently used