PDB renamed as outPDB
There should be two types of PDBs: outPDBs create PDBs and store the
result in a binary file; inPDBs load a PDB from a binary file and can be
used to guide heuristic search algorithms.

The class pdb is now a base class of these types of PDBs. This changeset
also includes an incomplete definition of outpdb which can be used to
generate PDBs but does not write information to a binary file yet.

inPDBs are not implemented yet